fix: extract Bash workdir validator
